+++ android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.gallery.app",
+    package = "net.gallery.bac",
 )
 
 android_resource(
     name = "res",
-    package = "com.gallery.app",
+    package = "net.gallery.bac",
     res = "src/main/res",
 )
