+++ android/app/release/output-metadata.json
@@ -4,15 +4,15 @@
     "type": "APK",
     "kind": "Directory"
   },
-  "applicationId": "com.gallery.app",
+  "applicationId": "net.gallery.bac",
   "variantName": "release",
   "elements": [
     {
       "type": "SINGLE",
       "filters": [],
       "attributes": [],
-      "versionCode": 1,
-      "versionName": "1.0",
+      "versionCode": 2,
+      "versionName": "1.8",
       "outputFile": "app-release.apk"
     }
   ],
