+++ android/app/src/main/java/net/gallery/bac/MainActivity.java
@@ -1,4 +1,4 @@
-package com.gallery.app;
+package net.gallery.bac;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -12,7 +12,7 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "gallery app";
+    return "gallerybac";
   }
 
   /**
