+++ android/app/src/main/java/net/gallery/bac/newarchitecture/components/MainComponentsRegistry.java
@@ -1,4 +1,4 @@
-package com.gallery.app.newarchitecture.components;
+package net.gallery.bac.newarchitecture.components;
 
 import com.facebook.jni.HybridData;
 import com.facebook.proguard.annotations.DoNotStrip;
