+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Gallery Bac</string>
+    <string name="app_name">gallery bac</string>
 </resources>
