+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "gallery app",
-  "displayName": "gallery app"
+  "name": "gallerybac",
+  "displayName": "gallerybac"
 }
