+++ ios/gallerybac.xcodeproj/xcshareddata/xcschemes/gallerybac.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "galleryapp.app"
-               BlueprintName = "galleryapp"
-               ReferencedContainer = "container:galleryapp.xcodeproj">
+               BuildableName = "gallerybac.app"
+               BlueprintName = "gallerybac"
+               ReferencedContainer = "container:gallerybac.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "galleryappTests.xctest"
-               BlueprintName = "galleryappTests"
-               ReferencedContainer = "container:galleryapp.xcodeproj">
+               BuildableName = "gallerybacTests.xctest"
+               BlueprintName = "gallerybacTests"
+               ReferencedContainer = "container:gallerybac.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "galleryapp.app"
-            BlueprintName = "galleryapp"
-            ReferencedContainer = "container:galleryapp.xcodeproj">
+            BuildableName = "gallerybac.app"
+            BlueprintName = "gallerybac"
+            ReferencedContainer = "container:gallerybac.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "galleryapp.app"
-            BlueprintName = "galleryapp"
-            ReferencedContainer = "container:galleryapp.xcodeproj">
+            BuildableName = "gallerybac.app"
+            BlueprintName = "gallerybac"
+            ReferencedContainer = "container:gallerybac.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
