+++ ios/gallerybac/AppDelegate.h
