+++ ios/gallerybac/AppDelegate.mm
@@ -41,7 +41,7 @@
   bridge.surfacePresenter = _bridgeAdapter.surfacePresenter;
 #endif
 
-  UIView *rootView = RCTAppSetupDefaultRootView(bridge, @"gallery app", nil);
+  UIView *rootView = RCTAppSetupDefaultRootView(bridge, @"gallerybac", nil);
 
   if (@available(iOS 13.0, *)) {
     rootView.backgroundColor = [UIColor systemBackgroundColor];
