+++ ios/gallerybac/Images.xcassets/AppIcon.appiconset/Contents.json
