+++ ios/gallerybac/Images.xcassets/Contents.json
