+++ ios/gallerybacTests/Info.plist
