+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "galleryapp",
+  "name": "gallerybac",
   "version": "0.0.1",
   "private": true,
   "scripts": {
