+++ src/components/Images.js
@@ -3,23 +3,46 @@ import { Text,Image, StyleSheet,View, TouchableOpacity,Modal } from 'react-nativ
 import ImageViewer from "react-native-image-zoom-viewer";
 import Ionicons from 'react-native-vector-icons/Ionicons';
 import RNFS from  'react-native-fs'
+import EditImage from './EditImage';
 //import Modal from "react-native-modal"
 
 function Images({path,uri,name,refreshing}) {
   const [imageSelect, setimageSelect] = useState([])
   const[openModal,setOpenModal] = useState(false)
+  const [openModalEdit, setOpenModalEdit]= useState(false)
   
   function toggleModal(){
     setimageSelect([{url:path,props:{}}])
     setOpenModal(!openModal)
   }
-
+  //Cortar la imagen
+  function handleCropImg(){
+    setOpenModalEdit(!openModal)
+  }
+  //Elimina la imagen
   function handleDeleteImg(){
-    RNFS.unlink(uri)
-    .then(()=>{console.log(`${name} se ha eliminado`); setOpenModal(!openModal);refreshing(true) })
-    .catch((err)=>{
-      console.log(err)
-    })
+    const filePath = path.split('////').pop()
+    RNFS.exists(filePath)
+    .then((res)=>{
+      if (res){
+       return RNFS.unlink(filePath)
+       .then(() => {
+         console.log('deleted');
+         RNFS.scanFile(filePath)
+           .then(() => {
+             console.log('scanned');
+             refreshing(true);
+             setOpenModal(!openModal)
+           })
+           .catch(err => {
+             console.log(err);
+           });
+       })
+       .catch((err) => {         
+           console.log(err);
+       })
+      }
+    }).catch(()=>console.log('no existe'))
   }
 
   return (
@@ -35,13 +58,14 @@ function Images({path,uri,name,refreshing}) {
               </TouchableOpacity>
              <ImageViewer style={styles.imageOpen} imageUrls={imageSelect}/>
              <View style={{flexDirection:'row',alignItems:'center', justifyContent:'space-between',marginHorizontal:50}}>
-             <TouchableOpacity >
+             <TouchableOpacity onPress={handleCropImg} >
                   <Ionicons name="create-outline" color={"#FFF"}size={30} style={{marginBottom:8, marginRight:5}} />
               </TouchableOpacity>
               <TouchableOpacity onPress={handleDeleteImg}>
                   <Ionicons name="trash-outline" color={"#FFF"}size={30} style={{marginBottom:8, marginRight:5}} />
               </TouchableOpacity>
              </View>
+             
           </View>
         </Modal>
     </View>
@@ -69,6 +93,12 @@ const styles = StyleSheet.create({
     imagen:{
         width: 110,
         height:110
+    },
+    cropImg:{
+      backgroundColor:'#000',
+      alignItems:'center',
+      justifyContent:'center'
     }
+
 })
 export default Images
