+++ src/helpers/generarId.js
@@ -1,6 +1,6 @@
 export const GenerarId = () => {
   const ramdom = Math.random().toString(36).substring(2, 11); //generar id de manera aleatoria
   const fecha = Date.now();
-  const id = ramdom + '_'+ fecha;
+  const id = ramdom + '-'+ fecha;
   return id;
 };
